backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
langchain>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
pydantic-settings>=2.5.0,<3.0.0
psycopg[binary]>=3.2.0,<4.0.0
pgvector>=0.3.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
pypdf>=5.0.0,<6.0.0
python-docx>=1.1.0,<2.0.0
PyJWT>=2.9.0,<3.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.0.1,<4.1.0
alembic>=1.16.0,<2.0.0
pytest>=8.3.0,<9.0.0
pytest-cov>=5.0.0,<7.0.0
httpx>=0.27.0,<1.0.0
